refactor(zonal_runners): split into a package mirroring depstor_builders/
Closes T2-A from docs/superpowers/reviews/2026-05-23-repo-fresh-eyes.md.

src/gfv2_params/zonal_runners.py (638 LOC, one file) -> a 7-file package
under src/gfv2_params/zonal_runners/:

  __init__.py  re-exports + heavy imports + GDAL toggle + startup heartbeat
  zonal.py     run_zonal_batch
  soils.py     run_soils_batch + 2 private helpers
  lulc.py      run_lulc_batch
  ssflux.py    run_ssflux_batch
  weights.py   run_build_weights
  merge.py     run_merge

Mirrors the depstor_builders/ pattern from PR #72. Restores symmetry between
the two Part-2 pipeline halves: adding a new param family is now "add a
submodule" instead of "edit a 638-line module" -- the same shape as adding a
new depstor step.

Pure refactor: function bodies move verbatim except for one byte in
lulc.py: Path(__file__).parents[2] -> parents[3], because the file moved
one level deeper and the relative-crosswalk-path resolution needs to stay
anchored to the repo root. Behaviour preserved (sanity-checked).

Public import surface (scripts/derive_zonal_params.py +
tests/test_merge_params.py) unchanged. Startup heartbeat + GDAL exception
toggle fire exactly once on first package import (verified).
No new dependencies.

The dispatch-table move (_BATCH_RUNNERS -> BATCH_RUNNERS in __init__.py) is
the separate follow-up commit on this branch.

Docs: README.md and CLAUDE.md path references updated to the package path.
